backend-flask/requirements.txt
# Python dependencies
flask
flask-alchemy
flask-migrate
flask-wtf
flask-cors
python-dotenv

